+++ src/main/java/com/example/fisafroexpay/controller/BankAccountController.java
@@ -2,9 +2,8 @@ package com.example.fisafroexpay.controller;
 
 import com.example.fisafroexpay.dto.BankAccountRequest;
 import com.example.fisafroexpay.dto.BankAccountResponse;
-import com.example.fisafroexpay.scrap.ScrapService;
 import com.example.fisafroexpay.service.BankAccountService;
-import java.io.IOException;
+import com.example.fisafroexpay.service.ScrapService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -47,10 +46,5 @@ public class BankAccountController {
   public ResponseEntity<String> deleteBankAccount(@RequestParam Long bankAccountId) {
     bankAccountService.deleteBankAccount(bankAccountId);
     return ResponseEntity.ok("Bank Account Deleted");
-  }
-
-  @GetMapping("/scrap")
-  public void getScrap() throws IOException {
-    scrapService.scrap();
   }
 }
